moved helpers for allele_similarities and synthetic_data into own modules, added script for best synthetic smoothing coef which 0.01
