+++ experiments/common.py
@@ -1,0 +1,28 @@
+#!/usr/bin/env python
+#
+# Copyright (c) 2015. Mount Sinai School of Medicine
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+from collections import defaultdict
+
+
+def curry_dictionary(key_pair_dict, default_value=0.0):
+    """
+    Transform dictionary from pairs of keys to dict -> dict -> float
+    """
+    result = defaultdict(dict)
+    for (a, b), value in key_pair_dict.items():
+        result[a][b] = value
+    return result
